+++ test-project-ts/moduleSuffixes/moduleSuffixes.ts
@@ -1,0 +1,5 @@
+import { usedFromNotExt } from './from';
+import { usedFromNative } from './use';
+import { usedFromTs } from './plain';
+
+const sum = usedFromNotExt + usedFromNative + usedFromTs;